Basic discord bot to demonstrate feasibility
The user types in !dex list and the Snapdex discord bot returns an unordered list of a preset set.
I started to add functionality to add to the dex but realised this would take me time and isn't part of this Sprint!
This version has been tested in TestBeef, MJ's private discord server. We should test on another server.